chore(must-define-example-schema): Weaken strictness of rule to only require examples for DTOs
Before a example was required for every key "content" which was quiet strict as examples can also be placed at attribute level. Now the rule only enforces a top-level example for every global DTO, thereby also matching the error message. It would be possible to enforce the usage of only DTOs for request bodys and responses in the future to further improve this rule.